Implementa exclusão de lista de presença Codaf
AB#140028
