+++ SME.ConectaFormacao.Aplicacao/Interfaces/Codaf/ICasoDeUsoExcluirCodafListaPresenca.cs
@@ -1,0 +1,9 @@
+﻿using SME.ConectaFormacao.Dominio.Comum;
+
+namespace SME.ConectaFormacao.Aplicacao.Interfaces.Codaf
+{
+    public interface ICasoDeUsoExcluirCodafListaPresenca
+    {
+        Task<Resultado> ExecutarAsync(long codafListaPresencaId);
+    }
+}
