+++ SME.ConectaFormacao.Infra.Dados/Repositorios/Interfaces/IRepositorioCodafListaPresenca.cs
@@ -10,5 +10,6 @@ namespace SME.ConectaFormacao.Infra.Dados.Repositorios.Interfaces
         Task<bool> TurmaJaTemListaDePresencaAsync(long propostaTurmaId, long listaPresencaId = 0);
         Task<ResultadoPaginado<ListagemResultadoCodafListaPresencaDto>> ObterListagemResultadoCodafListaPresencaPorFiltroAsync(FiltroListagemResultadoCodafListaPresencaDto filtro);
         Task<CodafListaPresenca?> ObterPorIdDetalhadoAsync(long id);
+        Task ExcluirAsync(long id);
     }
 }
